4. Add DemoRepository and DemoConstants
