+++ app/src/main/java/com/kuanhsien/app/sample/android_mvvm_demo/data/DemoConstants.java
@@ -1,0 +1,9 @@
+package com.kuanhsien.app.sample.android_mvvm_demo.data;
+
+public class DemoConstants {
+
+    // Demo Data
+    public static final String COVER = "cover";
+    public static final String AUTHOR = "author";
+
+}
