+++ app/src/main/java/com/kuanhsien/app/sample/android_mvvm_demo/data/InfoRepository.java
@@ -1,9 +1,7 @@
 package com.kuanhsien.app.sample.android_mvvm_demo.data;
 
 import android.content.Context;
-import android.os.AsyncTask;
 import java.util.List;
-
 import androidx.lifecycle.LiveData;
 
 public class InfoRepository {
@@ -36,19 +34,4 @@ public class InfoRepository {
     public void deleteInfo(InfoModel info) {
         mInfoDao.deleteInfo(info);
     }
-
-
-    public void prepareTestInfo() {
-        AsyncTask.execute(new Runnable() {
-
-            @Override
-            public void run() {
-                insertInfoAll(
-                        new InfoModel("Author", "A" , "Desc"),
-                        new InfoModel("Cover", "Cover-Title" , "Cover-Desc")
-                );
-            }
-        });
-    }
-
 }
