+++ app/src/main/res/layout/activity_item_detail.xml
@@ -66,7 +66,7 @@
         android:textStyle="normal"
         android:textColor="@android:color/white"
         android:gravity="center"
-        android:text="@string/btn_start"
+        tools:text="@string/btn_start"
         app:layout_constraintTop_toBottomOf="@+id/tv_item_detail_desc"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
